# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: ShopChain API
  version: 1.0.0
  description: |
    API de catálogo, inventario y pedidos de ShopChain.

    Autenticación con JWT Bearer, sin cookies ni sesión HTTP.
    POST /auth/login recibe JSON con identifier y password y devuelve accessToken,
    tokenType, expiresIn (segundos) y user. Enviar Authorization: Bearer <accessToken>.
    En Swagger, ejecutar login, copiar accessToken y pegarlo en Authorize.
    El token vence a los 30 minutos de su emisión por defecto; no se renueva.
    Al vencer, iniciar sesión nuevamente. Cerrar sesión significa borrar el token
    del cliente; no existe un endpoint de logout ni revocación de tokens.
    Los permisos del token se mantienen hasta su vencimiento, aunque se cambie
    el rol, correo, contraseña o estado del usuario en la base de datos.

    Todas las rutas de negocio requieren un JWT válido. Los permisos adicionales se
    indican en cada operación. Los listados son arrays sin paginación.
    Los IDs de los ejemplos son ilustrativos; consultar primero los listados.
    Las propiedades desconocidas de las solicitudes JSON se rechazan.
servers:
  - url: .
    description: Misma instancia y contexto que sirve /api/openapi.yml
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Autenticación]
      operationId: login
      summary: Iniciar sesión
      description: El correo identifica al usuario. Devuelve un JWT firmado con HS256 y duración fija.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [identifier, password]
              properties:
                identifier: {type: string, format: email, example: admin@shopchain.pe}
                password: {type: string, format: password, writeOnly: true}
      responses:
        '200':
          description: Token y datos del usuario autenticado
          headers:
            Cache-Control:
              schema: {type: string, example: no-store}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LoginResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /auth/me:
    get:
      tags: [Autenticación]
      operationId: currentUser
      summary: Consultar usuario identificado por el JWT
      responses:
        '200':
          description: Usuario actual
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Pegar accessToken devuelto por POST /auth/login, sin el prefijo Bearer.
...
  responses:
    BadRequest:
      description: Validación, formato o valor de parámetro inválido
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Unauthorized:
      description: JWT ausente, inválido o vencido; credenciales incorrectas o usuario inactivo al autenticar
      content:
        application/json:
          schema: {$ref: '#/components/schemas/SecurityError'}
          example: {status: 401, detail: Debes iniciar sesión.}
...
    NotFound:
      description: Recurso inexistente
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
  schemas:
    LoginResponse:
      type: object
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken: {type: string, description: JWT firmado con HS256}
        tokenType: {type: string, enum: [Bearer]}
        expiresIn: {type: integer, format: int64, description: Duración del token en segundos, example: 1800}
        user: {$ref: '#/components/schemas/User'}
    User:
      type: object
      additionalProperties: false
      required: [names, email, role]
      properties:
        id: {type: integer, format: int64, readOnly: true}
        names: {type: string, minLength: 1, maxLength: 120, example: Ana Torres}
        email: {type: string, format: email, maxLength: 254, example: ana@shopchain.pe}
        password:
          type: string
          format: password
          writeOnly: true
          minLength: 8
          maxLength: 72
          description: Obligatoria al crear; omitir al actualizar conserva la actual. Máximo 72 bytes UTF-8.
        role: {type: string, enum: [ADMINISTRADOR, ALMACÉN, TIENDA]}
        active: {type: boolean, default: true}
...
    SecurityError:
      type: object
      required: [status, detail]
      properties:
        status: {type: integer}
        detail: {type: string}
    Problem:
      type: object
      required: [status, detail]
      properties:
        type: {type: string, example: 'about:blank'}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string}
        fields:
          type: object
          description: Mensajes de validación por campo
          additionalProperties: {type: string}